Function zwave_command_handler::zwave_command_handler_register_handler

pub unsafe extern "C" fn zwave_command_handler_register_handler(
    new_command_class_handler: zwave_command_handler_t
) -> sl_status_t

@brief Register a command class handler.

This API replaces the legacy REGISTER_HANDLER() macro. Command handlers should perform their own initialization and register themselves to the command handler framework.

There is no sequence/order requirement, meaning that this function can be invoked before the Command Class handler framework is initialized.

@param new_command_class_handler A struct that contains the zwave_command_handler_t info that will be used for registering a Command Class handler.

Command class handlers can be fully disabled by setting both support and control to false. @returns SL_STATUS_OK If the Command Class handler was successfully registered @returns SL_STATUS_FAIL Otherwise